+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * Add support for timeouts
 * Make fields of pool status public
+* Fix possible deadlock and make implementation a lot simpler by using
+  the new tokio `Semaphore` and `Receiver::try_recv`.
 
 ## v0.3.0